move cairovm to separate package and refactor magic numbers into variables
