+++ internal/cairovm/vm_test.go
@@ -1,4 +1,4 @@
-package services
+package cairovm
 
 import (
 	"context"
